app/requirements.txt
azure-identity >= 1.19.0
azure-monitor-ingestion >= 1.0.4
azure-core >= 1.32.0
envparse >= 0.2.0